+++ src/table.rs
@@ -1,9 +1,9 @@
 use crate::db::TransactionGuard;
 use crate::sealed::Sealed;
 use crate::tree_store::{
-    AccessGuardMutInPlace, Btree, BtreeExtractIf, BtreeHeader, BtreeMut, BtreeRangeIter,
-    MAX_PAIR_LENGTH, MAX_VALUE_LENGTH, PageHint, PageNumber, PageTrackerPolicy, RawBtree,
-    TransactionalMemory,
+    AccessGuardMutInPlace, AllocationPolicy, Btree, BtreeExtractIf, BtreeHeader, BtreeMut,
+    BtreeRangeIter, MAX_PAIR_LENGTH, MAX_VALUE_LENGTH, PageHint, PageNumber, PageTrackerPolicy,
+    RawBtree, TransactionalMemory,
 };
 use crate::types::{Key, MutInPlaceValue, Value};
 use crate::{AccessGuard, AccessGuardMut, StorageError, WriteTransaction};
@@ -79,6 +79,7 @@ impl<'txn, K: Key + 'static, V: Value + 'static> Table<'txn, K, V> {
         allocated_pages: Arc<Mutex<PageTrackerPolicy>>,
         mem: Arc<TransactionalMemory>,
         transaction: &'txn WriteTransaction,
+        allocation_policy: AllocationPolicy,
     ) -> Table<'txn, K, V> {
         Table {
             name: name.to_string(),
@@ -89,6 +90,7 @@ impl<'txn, K: Key + 'static, V: Value + 'static> Table<'txn, K, V> {
                 mem,
                 freed_pages,
                 allocated_pages,
+                allocation_policy,
             ),
         }
     }